incorporating qt logger
